Add files via upload
Adds a globalScore singleton, and a McGuffin for the player to destroy on the level. It can be destroyed by calling the Destroy() function on it, (it can be found on physics layer 4 by an Area2D / Raycast or as part of groupMcGuffin.) When the first McGuffin is destroyed, the chaos timer in globalScore starts.